ruleguard: remove matchData allocations
Do not use interface type for matchData since it caused
heap allocations in both match and reject cases.
