+++ ruleguard/match_data.go
@@ -6,41 +6,14 @@ import (
 	"github.com/quasilyte/gogrep"
 )
 
-// matchData is used to handle both regexp and AST match sets in the same way.
-type matchData interface {
-	// TODO: don't use gogrep.CapturedNode type here.
-
-	Node() ast.Node
-	CaptureList() []gogrep.CapturedNode
-	CapturedByName(name string) (ast.Node, bool)
-}
-
-type commentMatchData struct {
-	node    ast.Node
-	capture []gogrep.CapturedNode
-}
-
-func (m commentMatchData) Node() ast.Node { return m.node }
-
-func (m commentMatchData) CaptureList() []gogrep.CapturedNode { return m.capture }
-
-func (m commentMatchData) CapturedByName(name string) (ast.Node, bool) {
-	for _, c := range m.capture {
-		if c.Name == name {
-			return c.Node, true
-		}
-	}
-	return nil, false
-}
-
-type astMatchData struct {
+type matchData struct {
 	match gogrep.MatchData
 }
 
-func (m astMatchData) Node() ast.Node { return m.match.Node }
+func (m matchData) Node() ast.Node { return m.match.Node }
 
-func (m astMatchData) CaptureList() []gogrep.CapturedNode { return m.match.Capture }
+func (m matchData) CaptureList() []gogrep.CapturedNode { return m.match.Capture }
 
-func (m astMatchData) CapturedByName(name string) (ast.Node, bool) {
+func (m matchData) CapturedByName(name string) (ast.Node, bool) {
 	return m.match.CapturedByName(name)
 }
